The new `test.py` introduces a fixture, that tracks the numerical error.

```sh
+----+-------------+------+----------+---------------------+---------------------+---------------------+---------------------+---------------------+-----------------+
|    |   Metric    | NDim |  DType   |   Baseline Error    |    SimSIMD Error    |  Accurate Duration  |  Baseline Duration  |  SimSIMD Duration   | SimSIMD Speedup |
+----+-------------+------+----------+---------------------+---------------------+---------------------+---------------------+---------------------+-----------------+
| 0  |   cosine    | 1536 | bfloat16 | 1.82e-09 ± 1.54e-09 | 2.68e-05 ± 2.60e-05 | 2.64e+04 ± 1.53e+03 | 2.69e+04 ± 7.08e+02 | 2.78e+03 ± 3.80e+02 |  9.84x ± 1.39x  |
| 1  |   cosine    | 1536 | float16  | 8.66e-06 ± 7.93e-06 | 6.39e-05 ± 4.71e-05 | 2.52e+04 ± 1.88e+03 | 5.41e+04 ± 2.03e+03 | 1.35e+03 ± 1.89e+02 | 40.91x ± 5.23x  |
| 2  |   cosine    | 1536 | float32  | 2.54e-09 ± 2.02e-09 | 2.66e-05 ± 3.75e-05 | 2.47e+04 ± 1.22e+03 | 2.68e+04 ± 3.49e+03 | 1.58e+03 ± 3.25e+02 | 17.65x ± 3.59x  |
| 3  |   cosine    | 1536 | float64  | 0.00e+00 ± 0.00e+00 | 2.54e-05 ± 2.31e-05 | 2.66e+04 ± 5.22e+03 | 1.93e+04 ± 1.62e+03 | 2.30e+03 ± 5.25e+02 |  8.76x ± 1.86x  |
| 4  |   cosine    | 1536 |   int8   | 0.00e+00 ± 0.00e+00 | 2.62e-05 ± 3.49e-05 | 1.90e+04 ± 6.51e+02 | 2.17e+04 ± 1.35e+03 | 1.67e+03 ± 2.38e+02 | 13.20x ± 2.02x  |
| 5  |     dot     | 1536 | float32c | 3.43e-07 ± 4.32e-07 | 4.60e-07 ± 3.87e-07 | 2.57e+03 ± 2.23e+03 | 1.47e+03 ± 1.18e+03 | 1.54e+03 ± 2.07e+02 |  0.93x ± 0.48x  |
| 6  |     dot     | 1536 | float64c | 0.00e+00 ± 0.00e+00 | 1.45e-15 ± 1.27e-15 | 2.36e+03 ± 9.89e+02 | 1.52e+03 ± 1.46e+02 | 2.01e+03 ± 2.31e+02 |  0.76x ± 0.10x  |
| 7  |   hamming   | 1536 |   bits   | 0.00e+00 ± 0.00e+00 | 0.00e+00 ± 0.00e+00 | 1.18e+04 ± 1.44e+03 | 7.62e+03 ± 5.74e+02 | 2.26e+03 ± 3.84e+02 |  3.48x ± 0.66x  |
| 8  |    inner    | 1536 | bfloat16 | 2.93e-07 ± 3.58e-07 | 5.00e-07 ± 9.64e-07 | 2.82e+03 ± 2.23e+02 | 1.81e+03 ± 1.81e+03 | 2.76e+03 ± 3.86e+02 |  0.65x ± 0.53x  |
| 9  |    inner    | 1536 | float16  | 1.65e-04 ± 1.10e-04 | 1.58e-02 ± 3.55e-02 | 2.12e+03 ± 2.42e+02 | 6.48e+03 ± 8.52e+02 | 1.21e+03 ± 1.79e+02 |  5.42x ± 0.91x  |
| 10 |    inner    | 1536 | float32  | 7.65e-07 ± 1.75e-06 | 1.12e-06 ± 3.51e-06 | 2.00e+03 ± 1.47e+02 | 1.34e+03 ± 1.74e+02 | 1.18e+03 ± 1.63e+02 |  1.15x ± 0.18x  |
| 11 |    inner    | 1536 | float64  | 0.00e+00 ± 0.00e+00 | 1.38e-15 ± 1.26e-15 | 2.30e+03 ± 8.06e+02 | 1.40e+03 ± 1.35e+02 | 1.94e+03 ± 4.76e+02 |  0.74x ± 0.09x  |
| 12 |    inner    | 1536 |   int8   | 0.00e+00 ± 0.00e+00 | 0.00e+00 ± 0.00e+00 | 2.07e+03 ± 2.42e+02 | 1.98e+03 ± 2.39e+02 | 1.74e+03 ± 3.13e+02 |  1.17x ± 0.25x  |
| 13 |   jaccard   | 1536 |   bits   | 0.00e+00 ± 0.00e+00 | 4.31e-17 ± 7.28e-17 | 1.69e+04 ± 4.24e+03 | 1.17e+04 ± 1.11e+03 | 2.26e+03 ± 3.99e+02 |  5.32x ± 1.06x  |
| 14 | sqeuclidean | 1536 | bfloat16 | 3.56e-07 ± 7.65e-08 | 3.52e-07 ± 7.67e-08 | 8.17e+03 ± 5.17e+02 | 4.78e+03 ± 1.98e+02 | 2.75e+03 ± 3.75e+02 |  1.78x ± 0.27x  |
| 15 | sqeuclidean | 1536 | float16  | 1.56e-04 ± 1.03e-04 | 9.54e-04 ± 6.12e-04 | 7.19e+03 ± 6.61e+02 | 1.61e+04 ± 1.49e+03 | 1.22e+03 ± 1.12e+02 | 13.22x ± 1.27x  |
| 16 | sqeuclidean | 1536 | float32  | 7.12e-08 ± 5.63e-08 | 1.09e-07 ± 7.97e-08 | 7.12e+03 ± 5.68e+02 | 4.71e+03 ± 4.29e+02 | 1.38e+03 ± 3.21e+02 |  3.55x ± 0.66x  |
| 17 | sqeuclidean | 1536 | float64  | 0.00e+00 ± 0.00e+00 | 5.26e-16 ± 4.39e-16 | 7.91e+03 ± 1.79e+03 | 5.10e+03 ± 6.22e+02 | 2.03e+03 ± 5.34e+02 |  2.66x ± 0.64x  |
| 18 | sqeuclidean | 1536 |   int8   | 0.00e+00 ± 0.00e+00 | 0.00e+00 ± 0.00e+00 | 5.06e+03 ± 2.62e+02 | 6.67e+03 ± 3.50e+02 | 1.78e+03 ± 3.64e+02 |  3.92x ± 0.86x  |
| 19 |    vdot     | 1536 | float32c | 3.66e-07 ± 5.36e-07 | 5.34e-07 ± 4.65e-07 | 2.01e+03 ± 3.02e+02 | 1.27e+03 ± 1.33e+02 | 1.09e+03 ± 1.74e+02 |  1.18x ± 0.14x  |
| 20 |    vdot     | 1536 | float64c | 0.00e+00 ± 0.00e+00 | 1.25e-15 ± 8.08e-16 | 1.94e+03 ± 1.21e+02 | 1.50e+03 ± 1.41e+02 | 1.48e+03 ± 1.91e+02 |  1.03x ± 0.12x  |
+----+-------------+------+----------+---------------------+---------------------+---------------------+---------------------+---------------------+-----------------+
```